Give query classes persistent store IDs
This commit adds an ID property to the Query and AggregatedQueryLog classes that represents the entity ID in a persistent data store. It was a deliberate decision not to include this property up to now as this allowed us to remain non-commital in terms of int or bigint data types in the SQL identity column. However, not having an ID property meant these classes cannot be used properly in C# DataTypes and probably a hundred other places.